change batch handler when splitting loaddb objeect file to a std::function
